FamilyBackend/requirements.txt
django==4.1.7 # for django
pillow==9.3.0 # for imagefield

djangorestframework==3.14.0 # For REST API
djangorestframework-simplejwt==5.2.2


drf-yasg==1.21.5 # for documentation
celery==5.2.7 # for Running cron jobs
redis==4.5.4 # database for websocket messages and celery results

channels==4.0.0 # for websockets
channels-redis==4.0.0
daphne==4.0.0



python-decouple==3.8 # for getting access to environment variables

django-celery-beat==2.4.0 # for storing celery jobs on the db

asgiref==3.5.2 # for websocket channel
aioredis==2.0.1 # for connecting websocket to redis



# testing
pytest==7.2.1 # for testing
pytest-sugar==0.9.6 # make test output pretty
pytest-django==4.5.2 # django pytest support
factory-boy==3.2.1 # generate test data
coverage==7.1.0  # check test coverage

# code quality
flake8==6.0.0  # python link checker
flake8-isort==6.0.0  # checks import order
black==23.1.0  # code formatter
pylint-django==2.5.3  # django lint checker
pre-commit==3.0.3  # runs pre-commit checks


websockets==10.4 # for websocket client

